chore: create `PrestoBaseEngineSpec` class that share common functions between Presto and Trino
