+++ superset/db_engine_specs/hive.py
@@ -49,7 +49,6 @@ if TYPE_CHECKING:
     # prevent circular imports
     from superset.models.core import Database
 
-
 logger = logging.getLogger(__name__)
 
 
@@ -261,10 +260,6 @@ class HiveEngineSpec(PrestoEngineSpec):
             return f"""CAST('{dttm
                 .isoformat(sep=" ", timespec="microseconds")}' AS TIMESTAMP)"""
         return None
-
-    @classmethod
-    def epoch_to_dttm(cls) -> str:
-        return "from_unixtime({col})"
 
     @classmethod
     def adjust_database_uri(
